tests: refactor and add more tests
